// modules/dashboard/jobLists/jobPost/enums/JobPostEnums.tsx
export enum ServiceTypes {
  BASIC_LISTING = 1,
  STAND_OUT_LISTING = 2,
  STAND_OUT_PREMIUM = 3,
}

export enum ResumeReceivingOptions {
  EMAIL = 1,
  HARD_COPY = 2,
  WALK_IN_INTERVIEW = 3,
}

export enum EmploymentStatus {
  FULL_TIME = 1,
  PART_TIME = 2,
  APPRENTICESHIP = 3,
  CONTRACTUAL = 4,
  // FREELANCE = 5,
}

export enum JobLevel {
  ENTRY = 1,
  MID = 2,
  TOP = 3,
}

export enum SalaryTypes {
  DAILY = 1,
  WEEKLY = 2,
  MONTHLY = 3,
  ANNUAL = 4,
}

export enum SalaryShowOption {
  YES = 1,
  NEGOTIABLE = 2,
}

export enum OtherBenefit {
  YES = 1,
  NO = 0,
}

export enum WorkPlaceTypes {
  HOME = 1,
  OFFICE = 2,
}

export enum JobPlaceTypes {
  INSIDE_BANGLADESH = 1,
  OUTSIDE_BANGLADESH = 2,
}

export enum SalaryReviewType {
  HALF_YEARLY = 1,
  YEARLY = 2,
}

export enum LunchFacilityType {
  PARTIALLY_SUBSIDIZE = 1,
  FULL_SUBSIDIZE = 2,
}

export enum Gender {
  MALE = 1,
  FEMALE = 2,
  OTHERS = 3,
}

export const SHOW = 1;